Verbesserte Label-API, Caching & Debug-Bereinigung
Die Debug-Logging-Funktion und zugehörige Aufrufe wurden entfernt. Die Variable _focus_index ist nun vom Typ int. Das Label long_text ersetzt c im Hauptprogramm und nutzt jetzt .config(wraplength = 50). Die alte deutsche Version und direkte _wraplength-Zuweisung wurden entfernt. Labels verwenden nun einen Cache für Funktions-Token, was die Effizienz beim Rendern erhöht. Die apply-Methode akzeptiert jetzt wraplength als Keyword-Argument, Fehlerbehandlung für ungültige Argumente wurde verbessert. Die Konstanten kLabelAutoWrap und das wrap-Keyword wurden entfernt. Die Methode get_fn_result_cache wurde für den Funktions-Token-Cache ergänzt.

+++ Source/Widgets/Container/ctui_container.h
@@ -6,7 +6,7 @@
 
 namespace ctui {
 	struct Container : Widget, protected defaults::WidgetDefaults {
-		size_t _focus_index = kFocusIndex;
+		int _focus_index = kFocusIndex;
 		std::vector<Widget*> _children = {};
 		
 		void make_child(Widget* child);
